requirements.txt
requests
feedparser
newspaper3k
lxml[html_clean]
google-genai